requirements.txt
Django==4.0
channels==3.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2